Create d3project.html
So this was my first big project, a bar chart in d3.js that takes data from an api and displays it nicely with a formatted tooltip. As a bonus, you can also change the color and length of the graph with the sliders at the top. I did think starting out that d3 would be good for big dataset visualization, however the thing I am most excited about is the capabilities d3 brings to IoT interfaces/controllers. Not only would you be able to check current sensor readings in a beautiful, visual way, but you would also have the capability to view historical data and recent trends (daily/last week data), all because d3 uses APIs. With this project, I have quite an appreciation now for APIs, given you can just update them whenever you want, and anything tied to that api (like a d3 visualization) will update each time the api is updated and the page is refreshed.